Don't delete .git directory

+++ run.py
@@ -82,10 +82,12 @@ def delete_directory(path: Path) -> None:
   trace('delete directory {}', path)
   shutil.rmtree(path, ignore_errors=True)
 
-def delete_contents(path: Path) -> None:
+def delete_contents(path: Path, excludes: set[str]) -> None:
   """Delete all entries from the given directory."""
   trace('delete directory contents {}', path)
   for entry in path.iterdir():
+    if entry in excludes:
+      continue
     if entry.is_symlink() or entry.is_file():
       entry.unlink(missing_ok=True)
     elif entry.is_dir():
@@ -289,7 +291,7 @@ def publish_docs() -> None:
   with temporary_directory(prefix='publish-docs') as tmpdir:
     copy(fs.docs / '_build' / 'html', tmpdir)
     exec('git', 'checkout', 'gh-pages')
-    delete_contents(fs.cwd)
+    delete_contents(fs.cwd, excludes=set(['.git']))
     copy(tmpdir, fs.cwd)
   exec('git', 'add', '.')
   exec('git', 'commit', '-m', 'Update gh-pages')
